Reorganize audit board sign-in container.
